src/main/java/com/tinkerpop/gremlin/compiler/functions/g/string/MatchesFunction.java
package com.tinkerpop.gremlin.compiler.functions.g.string;

import com.tinkerpop.gremlin.compiler.types.Atom;
import com.tinkerpop.gremlin.compiler.context.GremlinScriptContext;
import com.tinkerpop.gremlin.compiler.functions.AbstractFunction;
import com.tinkerpop.gremlin.compiler.operations.Operation;

import java.util.List;

/**
 * @author Marko A. Rodriguez (http://markorodriguez.com)
 */
public class MatchesFunction extends AbstractFunction<Boolean> {

    private static final String FUNCTION_NAME = "matches";


    public Atom<Boolean> compute(final List<Operation> parameters, final GremlinScriptContext context) throws RuntimeException {
        if (parameters.size() == 2) {
            final String string = (String) parameters.get(0).compute().getValue();
            final String regex = (String) parameters.get(1).compute().getValue();
            return new Atom<Boolean>(string.matches(regex));
        } else {
            throw new RuntimeException(this.createUnsupportedArgumentMessage());
        }
    }

    public String getFunctionName() {
        return FUNCTION_NAME;
    }
}